MyExceptions added. ProductDto -> ProductUpdateDto. The same convention kept for delete as for update or create. We will see.

+++ Kolisetka.Application/Features/Products/Handlers/Commands/DeleteProductCommandHandler.cs
@@ -1,4 +1,6 @@
 ﻿using AutoMapper;
+using Kolisetka.Application.DTOs.Validators;
+using Kolisetka.Application.Exceptions;
 using Kolisetka.Application.Features.Products.Requests.Commands;
 using Kolisetka.Application.Persistence.Contracts;
 using MediatR;
@@ -20,7 +22,15 @@ namespace Kolisetka.Application.Features.Products.Handlers.Commands
 
         public async Task<Unit> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
-            var product = await _productRepository.GetAsync(request.Id);
+            var validator = new ProductDeleteValidator(_productRepository);
+            var validationResult = await validator.ValidateAsync(request.ProductDto);
+
+            if (!validationResult.IsValid)
+                throw new ValidationException(validationResult);
+
+            var product = await _productRepository.GetAsync(request.ProductDto.Id);
+            _mapper.Map(request.ProductDto, product);
+
             await _productRepository.DeleteAsync(product);
 
             return Unit.Value;
